chore(.vscode): update typescript environment settings
One of the issues we face in the current setup is that it uses yarn scripts for in-IDE building for JS and TS, but it doesn't include the correct TS definitions.

This commit updates the user's IDE to point to the local path for installing TS lib definitions, which provides a consistent and deterministic environment for building and testing the project.