build: enable setting origin to environment variable
This commit adds the ability to set the ORIGIN environment variable to override the default origin used for API requests.

This is useful for local development or testing, where you may want to use a different origin than the production environment.

BREAKING CHANGE: The default origin is now https://roast.tfkhdyt.my.id instead of localhost:3000. If you are using a custom origin, you will need to set the ORIGIN environment variable to your custom origin.