fix: remove fork filter from repos list
This commit removes the filter that excludes forked repos from the list of repos displayed on the page.

The previous filter was causing issues because it was preventing repos from being displayed even if they were not actually forks. This change removes the filter so that all repos will be displayed, regardless of whether or not they are forked.