feat: add GitHub token to .env
This commit adds a GitHub token to the `.env` file, which is used to authenticate the application when making API requests to GitHub.

This change is necessary because the previous authentication method, which relied on a personal access token stored in the environment variables, was no longer working. The new method is more secure and will allow the application to continue to make API requests to GitHub.